server: redundant cast expressions removed, clean up

+++ src/server/sv_init.c
@@ -758,7 +758,7 @@ void SV_SpawnServer(char *server)
 
 	// get a new checksum feed and restart the file system
 	srand(Sys_Milliseconds());
-	sv.checksumFeed = (((int) rand() << 16) ^ rand()) ^ Sys_Milliseconds();
+	sv.checksumFeed = ((rand() << 16) ^ rand()) ^ Sys_Milliseconds();
 
 	// only comment out when you need a new pure checksum string and it's associated random feed
 	// Com_DPrintf("SV_SpawnServer checksum feed: %p\n", sv.checksumFeed);
